backend/pyproject.toml
[project]
name = "url-summarizer-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "httpx==0.28.1",
    "beautifulsoup4==4.12.3",
    "google-genai",
    "python-dotenv==1.0.1",
    "pydantic==2.10.4",
]

[dependency-groups]
dev = []
